+++ apps/admin-dashboard/app/routes/_dashboard.resume-books.create.tsx
@@ -7,13 +7,13 @@ import {
 import { Form as RemixForm, useActionData, useFetcher } from '@remix-run/react';
 import { useEffect } from 'react';
 
-import { createResumeBook } from '@oyster/core/resume-books';
+import { createResumeBook } from '@oyster/core/resumes';
 import {
   ResumeBookEndDateField,
   ResumeBookHiddenField,
   ResumeBookNameField,
   ResumeBookStartDateField,
-} from '@oyster/core/resume-books.ui';
+} from '@oyster/core/resumes.ui';
 import {
   Button,
   ComboboxPopover,
