+++ packages/core/src/member-profile.ui.ts
@@ -38,6 +38,6 @@ export {
   ChangePrimaryEmailInput,
   ListMembersInDirectoryWhere,
 } from './modules/member/member.types';
-export { CreateResumeBookInput } from './modules/resume-book/resume-book.types';
+export { CreateResumeBookInput } from './modules/resume/resume.types';
 export { ClaimSwagPackInput } from './modules/swag-pack/swag-pack.types';
 export { Environment, ListSearchParams } from './shared/types';
